Major changes on file imports
1. Converted paths to Pathlib object in paths_config.py
2. made some datatype changes in cpu_nms.pyx
3. Created 3D-GANTex notebook for seamless use

+++ FaceBoxes/utils/nms/cpu_nms.pyx
@@ -24,11 +24,11 @@ def cpu_nms(np.ndarray[np.float32_t, ndim=2] dets, np.float thresh):
     cdef np.ndarray[np.float32_t, ndim=1] scores = dets[:, 4]
 
     cdef np.ndarray[np.float32_t, ndim=1] areas = (x2 - x1 + 1) * (y2 - y1 + 1)
-    cdef np.ndarray[np.int32_t, ndim=1] order = scores.argsort()[::-1]
+    cdef np.ndarray[np.int64_t, ndim=1] order = scores.argsort()[::-1]
 
     cdef int ndets = dets.shape[0]
-    cdef np.ndarray[np.int32_t, ndim=1] suppressed = \
-            np.zeros((ndets), dtype=np.int32)
+    cdef np.ndarray[np.int64_t, ndim=1] suppressed = \
+            np.zeros((ndets), dtype=np.int64)
 
     # nominal indices
     cdef int _i, _j
